--- src/base/__tests__/Queuestream.test.ts
@@ -1,3 +1,0 @@
-describe('base/Queuestream', () => {
-  it.todo('test all methods');
-});
