--- src/constants.ts
@@ -1,3 +1,0 @@
-export const DEFAULTS = {
-  PREBUFFER_LENGTH: 12,
-};
