+++ src/providers/queuestream/queuestream.types.ts
@@ -1,0 +1,7 @@
+import type { Writable } from 'stream';
+
+export type Queuestream = {
+  getPrebuffer: () => Buffer;
+  currentPipe: (wrstr: Writable, opts?: { end?: boolean }) => Writable;
+  next: () => void;
+};
