+++ src/providers/tokens.ts
@@ -1,0 +1,5 @@
+import { createToken } from '@fridgefm/inverter';
+
+import type { Config } from './config';
+
+export const CONFIG_TOKEN = createToken<Config>('config');
