+++ src/providers/track/track.types.ts
@@ -1,0 +1,26 @@
+import type { Readable } from 'stream';
+import type { Tags } from 'node-id3';
+
+export type TrackPath = string;
+
+export type TrackStats = {
+  bitrate: number; // kbps
+  directory: string;
+  duration: number; // millisec
+  format: string;
+  fullPath: string;
+  name: string;
+  size: number; // bytes
+  tagsSize: number; // bytes
+  stringified: string;
+};
+
+export interface ShallowTrackMeta extends Tags {
+  origin: 'id3' | 'fs';
+}
+
+export type TTrack = {
+  fsStats: TrackStats;
+  getSound: () => [Error | null, Readable];
+  getMetaAsync: () => Promise<ShallowTrackMeta>;
+};
