+++ src/types/public.d.ts
@@ -1,8 +1,8 @@
 import type { ClientRequest, ServerResponse } from 'http';
-import type { TrackList, ReorderCb } from '../base/Playlist/Playlist.types';
-import type { TEmitter } from '../features/EventBus/events';
+import type { TrackList, ReorderCb } from '../providers/playlist/playlist.types';
+import type { TEmitter } from '../providers/events/events.types';
 
-export interface TStation {
+export type TStation = {
   start(): void;
   addFolder(folder: string): void;
   next(): void;
@@ -10,4 +10,4 @@ export interface TStation {
   getPlaylist(): TrackList;
   connectListener(req: ClientRequest | undefined, res: ServerResponse, cb: () => void): void;
   on: TEmitter['on'];
-}
+};
